DROPME Add a CI workflow to test cross compilers
This CI workflow is not intended to be merged on the main branch as is.
Its purpose is to help review the cross compiler.

If it is deemed interesting to keep it for special cases (to run the
test from time to time, rather less regularly), it could be triggered on
explicit `workflow_dispatch` or on a specific label. In that case, it
might make sense to amend it, in particular the tests for the new cases
in `sak` could be dropped.